Podspec, move base58 string.
